ref: replace http method based routes files by feature based
